extract copyInputItems out of NodeExecutionFunctions
