+++ packages/core/src/__tests__/node-execute-functions.test.ts
@@ -10,12 +10,7 @@ import type {
 import nock from 'nock';
 
 import type { ExecutionLifecycleHooks } from '@/execution-engine/execution-lifecycle-hooks';
-import {
-	copyInputItems,
-	invokeAxios,
-	proxyRequestToAxios,
-	removeEmptyBody,
-} from '@/node-execute-functions';
+import { invokeAxios, proxyRequestToAxios, removeEmptyBody } from '@/node-execute-functions';
 
 describe('NodeExecuteFunctions', () => {
 	describe('proxyRequestToAxios', () => {
@@ -223,54 +218,6 @@ describe('NodeExecuteFunctions', () => {
 
 			expect(response.status).toBe(200);
 			expect(response.data).toEqual({ success: true });
-		});
-	});
-
-	describe('copyInputItems', () => {
-		it('should pick only selected properties', () => {
-			const output = copyInputItems(
-				[
-					{
-						json: {
-							a: 1,
-							b: true,
-							c: {},
-						},
-					},
-				],
-				['a'],
-			);
-			expect(output).toEqual([{ a: 1 }]);
-		});
-
-		it('should convert undefined to null', () => {
-			const output = copyInputItems(
-				[
-					{
-						json: {
-							a: undefined,
-						},
-					},
-				],
-				['a'],
-			);
-			expect(output).toEqual([{ a: null }]);
-		});
-
-		it('should clone objects', () => {
-			const input = {
-				a: { b: 5 },
-			};
-			const output = copyInputItems(
-				[
-					{
-						json: input,
-					},
-				],
-				['a'],
-			);
-			expect(output[0].a).toEqual(input.a);
-			expect(output[0].a === input.a).toEqual(false);
 		});
 	});
