+++ packages/core/src/execution-engine/node-execution-context/execute-context.ts
@@ -25,7 +25,6 @@ import {
 
 // eslint-disable-next-line import/no-cycle
 import {
-	copyInputItems,
 	normalizeItems,
 	getRequestHelperFunctions,
 	getSSHTunnelFunctions,
@@ -40,6 +39,7 @@ import {
 	detectBinaryEncoding,
 } from './utils/binary-helper-functions';
 import { constructExecutionMetaData } from './utils/construct-execution-metadata';
+import { copyInputItems } from './utils/copy-input-items';
 import { getDeduplicationHelperFunctions } from './utils/deduplication-helper-functions';
 import { getFileSystemHelperFunctions } from './utils/file-system-helper-functions';
 import { getInputConnectionData } from './utils/get-input-connection-data';
