+++ packages/core/src/node-execute-functions.ts
@@ -59,7 +59,6 @@ import {
 	NodeHelpers,
 	NodeOperationError,
 	NodeSslError,
-	deepCopy,
 	isObjectEmpty,
 	ExecutionBaseError,
 	jsonParse,
@@ -1476,24 +1475,6 @@ export const getSchedulingFunctions = (workflow: Workflow): SchedulingFunctions 
 			scheduledTaskManager.registerCron(workflow, cronExpression, onTick),
 	};
 };
-
-/**
- * Returns a copy of the items which only contains the json data and
- * of that only the defined properties
- */
-export function copyInputItems(items: INodeExecutionData[], properties: string[]): IDataObject[] {
-	return items.map((item) => {
-		const newItem: IDataObject = {};
-		for (const property of properties) {
-			if (item.json[property] === undefined) {
-				newItem[property] = null;
-			} else {
-				newItem[property] = deepCopy(item.json[property]);
-			}
-		}
-		return newItem;
-	});
-}
 
 /**
  * Returns the execute functions the poll nodes have access to.
